# swagger.yaml
openapi: 3.0.0
info:
  title: Social Media API
  description: API documentation for the social media platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Enter your JWT token in the format: Bearer <your-token>
        
        Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        
        Note: The token is obtained from the login response
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        full_name:
          type: string
          description: Full name
    Post:
      type: object
      properties:
        id:
          type: integer
          description: Post ID
        content:
          type: string
          description: Post content
        media_url:
          type: string
          description: URL to media content (optional)
        comments_enabled:
          type: boolean
          description: Whether comments are enabled
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        likes_count:
          type: integer
          description: Number of likes
        comments_count:
          type: integer
          description: Number of comments
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
    ScheduledPost:
      type: object
      properties:
        id:
          type: integer
          description: Scheduled post ID
        user_id:
          type: integer
          description: User ID
        content:
          type: string
          description: Post content
        media_url:
          type: string
          description: URL to media content (optional)
        comments_enabled:
          type: boolean
          description: Whether comments are enabled
        scheduled_at:
          type: string
          format: date-time
          description: Scheduled time for posting
        status:
          type: string
          enum: [pending, posted, cancelled]
          description: Current status of the scheduled post
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    ScheduledPostCreate:
      type: object
      required:
        - content
        - scheduled_at
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          description: Post content
        media_url:
          type: string
          description: URL to media content (optional)
        comments_enabled:
          type: boolean
          description: Whether comments are enabled
        scheduled_at:
          type: string
          format: date-time
          description: Scheduled time for posting
...
paths:
...
  /api/posts:
    post:
      summary: Create a new post
      tags: [Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: Post content
                media_url:
                  type: string
                  description: URL to media content (optional)
                comments_enabled:
                  type: boolean
                  description: Whether comments are enabled
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/posts/{post_id}/likes:
    post:
      summary: Like a post
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
          description: Post ID
      responses:
        '200':
          description: Post liked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Unlike a post
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
          description: Post ID
      responses:
        '200':
          description: Post unliked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/{post_id}/comments:
    post:
      summary: Add a comment to a post
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
          description: Post ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 500
                  description: Comment content
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Comment ID
                  content:
                    type: string
                    description: Comment content
                  user:
                    $ref: '#/components/schemas/User'
                  created_at:
                    type: string
                    format: date-time
                    description: Creation timestamp
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get comments for a post
      tags: [Posts]
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
          description: Post ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of comments per page
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        content:
                          type: string
                        user:
                          $ref: '#/components/schemas/User'
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                    description: Total number of comments
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/{post_id}/comments/{comment_id}:
    delete:
      summary: Delete a comment
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
          description: Post ID
        - in: path
          name: comment_id
          required: true
          schema:
            type: integer
          description: Comment ID
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/scheduled-posts:
    post:
      summary: Create a new scheduled post
      tags: [Scheduled Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledPostCreate'
      responses:
        '201':
          description: Scheduled post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledPost'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get user's scheduled posts
      tags: [Scheduled Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, posted, cancelled]
          description: Filter by post status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of posts per page
      responses:
        '200':
          description: Scheduled posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduledPost'
                  total:
                    type: integer
                    description: Total number of posts
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...